Refactor README + DEVELOPER for external contribs
Major refactor to the project README, DEVELOPER (contribution guide),
and resources/README file for improved readability and improved context
for any external developer new to the project.

The goal is to get this project ready for contributions from developers
outside of BitGo -- in order to achieve that, much more context needed
to be added to the project for newcomers.